requirements.txt
astroid==2.4.2
boto3==1.16.44
botocore==1.19.44
click==7.1.2
Flask==1.1.2
Flask-WTF==0.14.3
gunicorn==20.0.4
isort==5.6.4
itsdangerous==1.1.0
Jinja2==2.11.2
jmespath==0.10.0
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
pylint==2.6.0
python-dateutil==2.8.1
python-decouple==3.3
s3transfer==0.3.3
six==1.15.0
tabulate==0.8.7
toml==0.10.2
urllib3==1.26.2
Werkzeug==1.0.1
wrapt==1.12.1
WTForms==2.3.3
